+++ src/types/index.ts
@@ -40,3 +40,4 @@ export * from './modelRuntime';
 export * from './novelMemory';
 export * from './chapterVersion';
 export * from './promptRegistry';
+export * from './feedbackDataset';